fix 12-03-2023

+++ app/src/main/java/com/example/weather/OpenMeteo/OpenMeteoDTO.kt
@@ -25,7 +25,7 @@ class OpenMeteoCurrentWeatherDTO(
     @SerializedName("weathercode")
     val weathercode: Int,
     @SerializedName("windspeed")
-    val windspeed: Int,
+    val windspeed: Float,
     @SerializedName("winddirection")
     val winddirection: Int,
 )
